+++ pkgs/google_cloud_logging/test/traceparent_test.dart
@@ -14,6 +14,7 @@
 
 import 'dart:async';
 
+import 'package:google_cloud_logging/interop.dart';
 import 'package:google_cloud_logging/src/traceparent.dart';
 import 'package:test/test.dart';
 
@@ -133,9 +134,9 @@ void main() {
           });
         },
         zoneValues: {
-          'traceparent':
+          traceparentHeaderValueZoneVariable:
               '00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01',
-          'google_cloud_project': 'zone-project',
+          googleCloudProjectIdZoneVariable: 'zone-project',
         },
       );
     });
@@ -152,9 +153,9 @@ void main() {
           });
         },
         zoneValues: {
-          'traceparent':
+          traceparentHeaderValueZoneVariable:
               '00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01',
-          'google_cloud_project': 'zone-project',
+          googleCloudProjectIdZoneVariable: 'zone-project',
         },
       );
     });