+++ pkgs/google_cloud_storage/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.6.4-wip
+
+* Add an `ifMetagenerationNotMatch` parameter to `Storage.patchBucket`,
+  `Storage.uploadObject`, and `Storage.uploadObjectFromString`. If the
+  precondition is not satisfied, a `NotModifiedException` is thrown.
+
+## 0.6.3
+
+* Add `prefix`, `delimiter`, and `includeTrailingDelimiter` parameters to
+  `Storage.listObjects`.
+* Add `Storage.rewriteObject`.
+* Add `StorageObject.rewrite`.
+
 ## 0.6.2
 
 * Add retries to `uploadObjectFromSink`.